% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurr_Stats.R
\name{occurr_stats}
\alias{occurr_stats}
\title{Calculate the monthly occurrence statistic of a rainfall Time Series}
\usage{
occurr_stats(monthlydata)
}
\arguments{
\item{monthlydata}{The monthly Time Serie}
}
\value{
A dataframe containing 4 values: Probability of Dry-Wet Event "P_DW", Probability of Wet-Wet Event "PWW", Marginal Probability of Wet Days "Pie_W, Marginal Probability of Dry Days "Pie_D"
}
\description{
Calculate the monthly occurrence statistic of a rainfall Time Series
}
